[USERGRID-608] Added initial version of delete flow

+++ stack/core/src/main/java/org/apache/usergrid/corepersistence/asyncevents/EventBuilderImpl.java
@@ -115,6 +115,9 @@ public class EventBuilderImpl implements EventBuilder {
     }
 
 
+    //Does the queue entityDelete mark the entity then immediately does to the deleteEntityIndex. seems like
+    //it'll need to be pushed up higher so we can do the marking that isn't async or does it not matter?
+
     @Override
     public EntityDeleteResults queueEntityDelete( final ApplicationScope applicationScope, final Id entityId ) {
         log.debug( "Deleting entity id from index in app scope {} with entityId {} }", applicationScope, entityId );
@@ -125,19 +128,25 @@ public class EventBuilderImpl implements EventBuilder {
 
         //needs get versions here.
 
+
         //TODO: change this to be an observable
         //so we get these versions and loop through them until we find the MvccLogEntry that is marked as delete.
         //TODO: evauluate this to possibly be an observable to pass to the nextmethod.
         MvccLogEntry mostRecentlyMarked = ecm.getVersions( entityId ).toBlocking()
-                                             .first( mvccLogEntry -> mvccLogEntry.getState()== MvccLogEntry.State.DELETED );
+                                             .firstOrDefault( null,
+                                                 mvccLogEntry -> mvccLogEntry.getState() == MvccLogEntry.State.DELETED );
 
+        //If there is nothing marked then we shouldn't return any results.
+        //TODO: evaluate if we want to return null or return empty observable when we don't have any results marked as deleted.
+        if(mostRecentlyMarked == null)
+            return null;
 
         //observable of index operation messages
         //this method will need the most recent version.
         //When we go to compact the graph make sure you turn on the debugging mode for the deleted nodes so
         //we can verify that we mark them. That said that part seems kinda done. as we also delete the mvcc buffers.
         final Observable<IndexOperationMessage> edgeObservable =
-            indexService.deleteEntityIndexes( applicationScope, entityId,mostRecentlyMarked.getVersion() );
+            indexService.deleteEntityIndexes( applicationScope, entityId, mostRecentlyMarked.getVersion() );
 
 
         //observable of entries as the batches are deleted
